backend/pyproject.toml
[project]
name = "ascent-log-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "galexad",email = "gabrielaalexandruu@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
